Add function to search note form Lua code
Allow us to use lua code to export a bunch of notes based on the query
WHERE condition

Normal search in search window only allow 200 notes as maximum return.
And selec all notes there to export is hard (many clicks). This will
allow us to script this tasks and overcome that limitation.